Renamed to deploy & kube-deploy Make kube-deploy a command with up|down|push subcommands
