+++ hack/build-go.sh
@@ -54,7 +54,7 @@ if [[ ${#targets[@]} -eq 0 ]]; then
     cmd/controller-manager
     cmd/kubelet
     cmd/kubecfg
-    cmd/kube-up
+    cmd/kube-deploy
     plugin/cmd/scheduler
   )
 fi
